+++ tardsquad_discord_bot/ratingentry.py
@@ -4,8 +4,8 @@ RATING_EXPIRES_AFTER_H = 6
 
 
 class RatingEntry:
-    def __init__(self, user, rating, reason=None):
-        self.user = user
+    def __init__(self, username, rating, reason=None):
+        self.username = username
         self.rating = rating
         self.reason = reason
         self.expiry_time = datetime.now() + timedelta(hours=RATING_EXPIRES_AFTER_H)